move /v1/network/details-v2 into /v2/network/details
