+++ nym-api/src/network/handlers/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+pub mod v1;
+pub mod v2;
