+++ nym-api/src/support/http/router.rs
@@ -4,7 +4,6 @@
 use crate::circulating_supply_api::handlers::circulating_supply_routes;
 use crate::ecash::api_routes::handlers::ecash_routes;
 use crate::mixnet_contract_cache::handlers::{epoch_routes, legacy_nodes_routes};
-use crate::network::handlers::nym_network_routes;
 use crate::node_families::handlers as node_families_handlers;
 use crate::node_status_api::handlers::status_routes;
 use crate::support::http::openapi::ApiDoc;
@@ -13,7 +12,7 @@ use crate::unstable_routes::v1::unstable_routes_v1;
 use crate::unstable_routes::v2::unstable_routes_v2;
 use crate::unstable_routes::v3::unstable_routes_v3;
 use crate::utility_routes::utility_routes;
-use crate::{nym_nodes, status};
+use crate::{network, nym_nodes, status};
 use anyhow::anyhow;
 use axum::response::Redirect;
 use axum::routing::get;
@@ -47,7 +46,7 @@ impl RouterBuilder {
             .nest("/epoch", epoch_routes())
             .nest("/circulating-supply", circulating_supply_routes())
             .nest("/status", status_routes(network_monitor))
-            .nest("/network", nym_network_routes())
+            .nest("/network", network::handlers::v1::routes())
             .nest("/api-status", status::handlers::api_status_routes())
             .nest("/nym-nodes", nym_nodes::handlers::v1::routes())
             .nest("/node-families", node_families_handlers::routes())
@@ -66,6 +65,7 @@ impl RouterBuilder {
     fn v2_routes() -> Router<AppState> {
         Router::new()
             .nest("/unstable", unstable_routes_v2())
+            .nest("/network", network::handlers::v2::routes())
             .nest("/nym-nodes", nym_nodes::handlers::v2::routes())
     }
